+++ rsa.h
@@ -1,3 +1,5 @@
+#include <stdint.h>
+
 #ifndef EXTERN
 #define EXTERN extern
 #endif
@@ -11,15 +13,15 @@ EXTERN unsigned char A_num[KEYSIZE], B_num[KEYSIZE], C_num[KEYSIZE], N_num[KEYSI
 
 
 /*
- * function to get a long from a stream of bytes
+ * function to get a 32-bit integer from a stream of bytes
  */
-#define Getlong(x) ( ((long)(x)[0] << 24) | ((long)(x)[1] << 16) | ((long)(x)[2] << 8) | ((long)(x)[3]) )
+#define Getlong(x) ( ((int32_t)(x)[0] << 24) | ((int32_t)(x)[1] << 16) | ((int32_t)(x)[2] << 8) | ((int32_t)(x)[3]) )
 
 /*
- * function to put a long into a stream of bytes
+ * function to put a 32-bit integer into a stream of bytes
  */
 static void
-Putlong(unsigned char *dest, long x)
+Putlong(unsigned char *dest, int32_t x)
 {
 	dest[0] = (x >> 24) & 0x00ff;
 	dest[1] = (x >> 16) & 0x00ff;
